requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
greenlet==0.4.17
idna==2.10
itsdangerous==1.1.0
MarkupSafe==1.1.1
pycryptodomex==3.9.8
python-engineio==3.13.2
python-socketio==4.6.0
requests==2.24.0
six==1.15.0
urllib3==1.25.11
uuid==1.30
websocket-client==0.57.0
Werkzeug==1.0.1
